Replace .rodata.* -> .rodata* and .srodata.* -> .srodata* in flexpret.ld, which solves a lot of issues. Now, this looks the same as .text*. I suspect the reason this works is because the wildcard pattern .rodata.* does not actually match .rodata itself. And the same goes for .srodata.
Also update sections test in c-tests folder with more sections tested. To view the results of the changes to the linker script, try to remove the changes and run the sections test.

Furthermore, some multithreading tests failed before this change was made to the linker script. They seem to pass now. Try to run them all with and without the changes to the linker script.

Note that the bug described in issue 78 has not been solved yet.