Fix execution using multiple hw threads with bootloader
This includes:
1. Halting threads with tid != 0 until the bootloader has completed.
When the bootloader has completed, they jump to the applicaton
(just as they do in the usual startup.c without bootloader).
2. Disabling instruction memory protection before the bootloader
is run. The protection is enabled again for the entire IMEM after
the bootloader has completed. (This was an issue because FlexPRET
is configured to have no IMEM protection when there is 1 hw thread
but protection otherwise, which caused a crash.)

Has been tested on emulator with 2 and 4 hw threads.
Has been tested on FPGA with 4 hw threads. Execution appears to work
there but not printing.