Fix bug with long external interrupts; fix bug with external interrupts triggering the incorrect thread; minor changes elsewhere.
Bug with long external interrupts:
When an interrupt signal is high for a long time, it can possible trigger the same interrupt handler N times.
This is ok, but triggering the interrupt over again would cause it to jump back to the wrong PC. This has been fixed.

Bug with external interupt triggering the incorrect thread:
When an external interrupt was found, the thread currently in the execute stage would have it's interrupt
handler triggered. This is not correct, since external interrupt signal <N> should trigger thread <N>'s
interrupt handler.

Tests for both these cases have been added.

The minor changes are with the layout of the FPGA IO, and some simplifications to the bootloader.