Do partial port from make to cmake. FlexPRET builds with cmake now, but not yet software.
Changes are:
1. No longer pass Scala program a "configuration string" that builds
FlexPRET. This approach was not very scalable because the string
would get very long. Instead, use appliction.conf and third-party
library PureConfig to pass configuration.
2. Add a single point of truth for the configuration, from which
all other configuration is derived. The default configuration
is cmake/configs/default.cmake, but it can be overriden by user.
This file is used to generate config.h, config.ld, and
application.conf.
3. Port emulator.mk to cmake as well.